Extract database and weather function; Start extracting tables into separated classes
